requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
faiss-cpu
python-dotenv
unstructured
tiktoken
transformers
sentence-transformers